refactor widgets
