+++ editor_app/app.py
@@ -31,7 +31,6 @@ from json import dump as jdump
 from enum import Enum
 import requests
 
-from customwidgets import ColorSelectWidget, ToolBoxWidget
 import customwidgets as cwdg
 import widgets as wdg
 import auxiliaries as aux
@@ -99,53 +98,6 @@ class MainWindowUi(QMainWindow):
         self.button_pull_sprite = self.findChild(QToolButton, "toolButton_pullSprite")
         self.button_pull_sprite.clicked.connect(self.pullSprite)
 
-
-        #### Tools
-        widget_tool_box = self.findChild(QWidget, "widget_tool_box")
-        self.toolbox = ToolBoxWidget()
-        self.giveTool = self.toolbox.giveTool
-        self.giveBrush = self.toolbox.giveBrush
-        self.giveBrushsize = self.toolbox.giveBrushsize
-        self.giveAirbrushStrength = self.toolbox.giveAirbrushStrength
-
-        widget_tool_box.layout().addWidget(self.toolbox)
-
-
-        # Color Panel
-        self.widget_color_panel = self.findChild(QGroupBox, "groupBox_selectedColor")
-
-        self.color_select_panel = ColorSelectWidget(self.current_palette, True, True, True)
-        self.giveActiveShade = self.color_select_panel.giveActiveShade #this is a function wrapper to get the current active shade
-
-        self.widget_color_panel.layout().addWidget(self.color_select_panel)
-
-        # Color Manipulations
-        self.checkbox_all_views = self.findChild(
-            QCheckBox, "checkBox_allViews")
-
-        self.button_remap_to = self.findChild(
-            QPushButton, "pushButton_remapTo")
-        self.combobox_remap_to_color = self.findChild(
-            QComboBox, "comboBox_remapToColor")
-
-        self.combobox_remap_to_color.addItems(
-            list(self.current_palette.color_dict))
-        self.combobox_remap_to_color.setCurrentIndex(
-            self.current_palette.color_dict['1st Remap'])
-
-        self.button_incr_brightness = self.findChild(
-            QPushButton, "pushButton_incrBrightness")
-        self.button_decr_brightness = self.findChild(
-            QPushButton, "pushButton_decrBrightness")
-        self.button_remove_color = self.findChild(
-            QPushButton, "pushButton_deleteColor")
-
-        self.button_remap_to.clicked.connect(self.colorRemapTo)
-        self.button_incr_brightness.clicked.connect(lambda x, step = 1: self.colorChangeBrightness(step))
-        self.button_decr_brightness.clicked.connect(lambda x, step = -1: self.colorChangeBrightness(step))
-        self.button_remove_color.clicked.connect(self.colorRemove)
-
-
         #### Menubar
         self.actionSmallScenery.triggered.connect(lambda x: self.newObject(cts.Type.SMALL))
         self.actionOpenFile.triggered.connect(self.openObjectFile)
@@ -175,6 +127,19 @@ class MainWindowUi(QMainWindow):
         self.actionCheckForUpdates.triggered.connect(self.checkForUpdates)
         self.actionAbout.triggered.connect(self.aboutPage)
 
+        ### Left bar
+        container = self.container_left_bar.layout()
+
+        self.tool_widget = wdg.ToolWidgetSprite(self)
+        container.addWidget(self.tool_widget)
+
+        #function wrappers
+        self.giveTool = self.tool_widget.toolbox.giveTool
+        self.giveBrush = self.tool_widget.toolbox.giveBrush
+        self.giveBrushsize = self.tool_widget.toolbox.giveBrushsize
+        self.giveAirbrushStrength = self.tool_widget.toolbox.giveAirbrushStrength
+
+        self.giveActiveShade = self.tool_widget.color_select_panel.giveActiveShade
 
         #Load empty object if not started with objects
 
@@ -323,7 +288,7 @@ class MainWindowUi(QMainWindow):
             self.actionPaletteOld.setChecked(True)
 
         if update_widgets:
-            self.color_select_panel.switchPalette(self.current_palette)
+            self.tool_widget.color_select_panel.switchPalette(self.current_palette)
             for index in range(self.object_tabs.count()):
                 tab = self.object_tabs.widget(index)
                 tab.o.switchPalette(self.current_palette)
@@ -398,7 +363,7 @@ class MainWindowUi(QMainWindow):
             o.switchPalette(self.current_palette)
 
 
-        object_tab = wdg.ObjectTabSS(o, self, filepath, author_id = author_id)
+        object_tab = wdg.ObjectTab(o, self, filepath, author_id = author_id)
 
         sprite_tab = wdg.SpriteTab(self, object_tab)
 
@@ -433,13 +398,13 @@ class MainWindowUi(QMainWindow):
                 self.button_lock.setChecked(True)
                 self.button_pull_sprite.setEnabled(False)
                 self.button_push_sprite.setEnabled(False)
-                self.checkbox_all_views.setEnabled(True)
+                self.tool_widget.checkbox_all_views.setEnabled(True)
             else:
                 self.button_lock.setChecked(False)
                 self.button_pull_sprite.setEnabled(True)
                 self.button_push_sprite.setEnabled(True)
-                self.checkbox_all_views.setEnabled(False)
-                self.checkbox_all_views.setChecked(False)
+                self.tool_widget.checkbox_all_views.setEnabled(False)
+                self.tool_widget.checkbox_all_views.setChecked(False)
 
 
     def lockClicked(self):
@@ -467,7 +432,7 @@ class MainWindowUi(QMainWindow):
 
             self.button_pull_sprite.setEnabled(False)
             self.button_push_sprite.setEnabled(False)
-            self.checkbox_all_views.setEnabled(True)
+            self.tool_widget.checkbox_all_views.setEnabled(True)
 
         else:
 
@@ -481,8 +446,8 @@ class MainWindowUi(QMainWindow):
 
             self.button_pull_sprite.setEnabled(True)
             self.button_push_sprite.setEnabled(True)
-            self.checkbox_all_views.setEnabled(False)
-            self.checkbox_all_views.setChecked(False)
+            self.tool_widget.checkbox_all_views.setEnabled(False)
+            self.tool_widget.checkbox_all_views.setChecked(False)
 
 
     def pushSprite(self):
@@ -509,7 +474,7 @@ class MainWindowUi(QMainWindow):
         if not self.current_palette == pal.orct:
             o.switchPalette(self.current_palette)
 
-        object_tab = wdg.ObjectTabSS(o, self, author = self.settings['author'], author_id = self.settings['author_id'])
+        object_tab = wdg.ObjectTab(o, self, author = self.settings['author'], author_id = self.settings['author_id'])
         sprite_tab = wdg.SpriteTab(self, object_tab)
 
         object_tab.lockWithSpriteTab(sprite_tab)
@@ -603,49 +568,6 @@ class MainWindowUi(QMainWindow):
         msg.exec_()
 
 
-    #### Color manipulations
-
-    def colorRemapTo(self):
-        color_remap = self.combobox_remap_to_color.currentText()
-        selected_colors = self.color_select_panel.selectedColors()
-
-        if self.checkbox_all_views.isChecked():
-            widget = self.object_tabs.currentWidget()
-
-            widget.colorRemapToAll(color_remap, selected_colors)
-
-        else:
-            widget = self.sprite_tabs.currentWidget()
-
-            widget.colorRemap(color_remap, selected_colors)
-
-
-    def colorChangeBrightness(self, step):
-        selected_colors = self.color_select_panel.selectedColors()
-
-        if self.checkbox_all_views.isChecked():
-            widget = self.object_tabs.currentWidget()
-
-            widget.colorChangeBrightnessAll(step, selected_colors)
-
-        else:
-            widget = self.sprite_tabs.currentWidget()
-
-            widget.colorChangeBrightness(step, selected_colors)
-
-    def colorRemove(self):
-        selected_colors = self.color_select_panel.selectedColors()
-
-        if self.checkbox_all_views.isChecked():
-            widget = self.object_tabs.currentWidget()
-
-            widget.colorRemoveAll(selected_colors)
-
-        else:
-            widget = self.sprite_tabs.currentWidget()
-
-            widget.colorRemove(selected_colors)
-
     ### Sprite Actions
 
 
@@ -653,13 +575,13 @@ class MainWindowUi(QMainWindow):
 
     def keyPressEvent(self, e):
         if e.key() == QtCore.Qt.Key_Alt:
-            self.toolbox.selectTool(cwdg.Tools.EYEDROPPER)
+            self.tool_widget.toolbox.selectTool(cwdg.Tools.EYEDROPPER)
 
 
 
     def keyReleaseEvent(self, e):
         if e.key() == QtCore.Qt.Key_Alt:
-            self.toolbox.restoreTool()
+            self.tool_widget.toolbox.restoreTool()
 
     def dragEnterEvent(self, e):
         if e.mimeData().hasUrls:
